frontend/requirements.txt
spotipy
redis
pillow

psutil
python-mpd2
